Move functions to module_utils
