+++ plugins/modules/kubeconfig.py
@@ -305,10 +305,14 @@ dest:
   sample: /home/user/.kube/config
 """
 
-from ansible.module_utils.basic import AnsibleModule
-
+from ansible.module_utils.basic import AnsibleModule, missing_required_lib
+from ansible_collections.kubernetes.core.plugins.module_utils.kubeconfig import (
+    load_yaml_file,
+    merge_by_name,
+    hash_data,
+    write_file,
+)
 import os
-import hashlib
 import traceback
 
 try:
@@ -319,95 +323,6 @@ try:
 except ImportError:
     IMP_YAML = False
     IMP_YAML_ERR = traceback.format_exc()
-
-
-def load_yaml_file(path):
-    if not IMP_YAML:
-        raise Exception(f"PyYAML is required:\n{IMP_YAML_ERR}")
-
-    if not path or not os.path.exists(path):
-        return {}
-
-    with open(path, "r") as f:
-        return yaml.safe_load(f) or {}
-
-
-def deep_merge(base, updates):
-    result = base.copy()
-
-    for key, value in updates.items():
-        if key in result and isinstance(result[key], dict) and isinstance(value, dict):
-            result[key] = deep_merge(result[key], value)
-        else:
-            result[key] = value
-
-    return result
-
-
-def merge_by_name(existing, new):
-
-    # Build dict of existing resources keyed by name
-    merged = {}
-    for item in existing:
-        if isinstance(item, dict) and "name" in item:
-            merged[item["name"]] = item
-
-    # Process new resources
-    for item in new:
-        if not isinstance(item, dict) or "name" not in item:
-            continue
-
-        name = item["name"]
-        behavior = item.get("behavior", "merge")
-
-        item_copy = {k: v for k, v in item.items() if k != "behavior"}
-
-        if name in merged:
-            # Entry exists - apply behavior
-            if behavior == "keep":
-                continue
-            elif behavior == "replace":
-                merged[name] = item_copy
-            else:
-                result = {"name": name}
-
-                for key in ["cluster", "user", "context"]:
-                    if key in merged[name] or key in item_copy:
-                        existing_config = merged[name].get(key, {})
-                        new_config = item_copy.get(key, {})
-                        result[key] = deep_merge(existing_config, new_config)
-
-                for key in merged[name]:
-                    if key not in ["name", "cluster", "user", "context"]:
-                        result[key] = merged[name][key]
-
-                for key in item_copy:
-                    if (
-                        key not in ["name", "cluster", "user", "context"]
-                        and key not in result
-                    ):
-                        result[key] = item_copy[key]
-
-                merged[name] = result
-        else:
-            merged[name] = item_copy
-
-    return list(merged.values())
-
-
-def hash_data(data):
-    """Generate SHA-256 hash for idempotency checking."""
-    return hashlib.sha256(yaml.safe_dump(data, sort_keys=True).encode()).hexdigest()
-
-
-def write_file(dest, data):
-    if not dest:
-        return False
-
-    with open(dest, "w") as f:
-        yaml.safe_dump(data, f, sort_keys=False)
-
-    return True
 
 
 def run_module():
@@ -435,6 +350,8 @@ def run_module():
 
     # Load existing kubeconfig
     try:
+        if not IMP_YAML:
+            module.fail_json(msg=missing_required_lib("pyyaml"))
         existing = load_yaml_file(path) if path else {}
     except Exception as e:
         module.fail_json(msg=f"Failed to load existing kubeconfig: {str(e)}")
